Allow linking settings on comparison pages (#519)
Closes #504 

Mostly a refactor

Share the same link component on both article and comparison pages

Share the tests, but genericize them by base page
