+++ react/test/comparison_link_settings_test.ts
@@ -1,0 +1,3 @@
+import { linkSettingsTests } from './link_settings_test_template'
+
+linkSettingsTests('/comparison.html?longnames=%5B%22Knoxville+%5BUrban+Area%5D%2C+TN%2C+USA%22%2C%22Louisville+KY+HSA%2C+Louisville+KY+HRR%2C+USA%22%5D')
