+++ react/test/symlink_test.ts
@@ -8,29 +8,29 @@ function symlink_test(name: string, link: string, expected: string | undefined =
 
     test(name, async (t) => {
         await screencap(t)
-        await t.expect(getLocation()).eql(`${TARGET}${expected}`)
+        await t.expect(getLocation()).eql(`${TARGET}${expected}&s=3t2X5xvsKo`)
     })
 }
 
 symlink_test(
     'national',
     '/article.html?longname=Timor-Leste',
-    '/article.html?longname=East+Timor&s=3t2X5xvsKo',
+    '/article.html?longname=East+Timor',
 )
 symlink_test(
     'subnational',
     '/article.html?longname=Haut-Lomami%2C+Congo%2C+The+Democratic+Republic+of+the',
-    '/article.html?longname=Haut-Lomami%2C+Democratic+Republic+of+the+Congo&s=3t2X5xvsKo',
+    '/article.html?longname=Haut-Lomami%2C+Democratic+Republic+of+the+Congo',
 )
 symlink_test(
     'urban center',
     '/article.html?longname=Tehr%C4%81n%2C+Iran%2C+Islamic+Republic+of',
-    '/article.html?longname=Tehr%C4%81n%2C+Iran&s=3t2X5xvsKo',
+    '/article.html?longname=Tehr%C4%81n%2C+Iran',
 )
 symlink_test(
     '5mpc',
     '/article.html?longname=Pyongyang+5MPC%2C+Korea%2C+Democratic+People%27s+Republic+of',
-    '/article.html?longname=Pyongyang+5MPC%2C+North+Korea&s=3t2X5xvsKo',
+    '/article.html?longname=Pyongyang+5MPC%2C+North+Korea',
 )
 
 symlink_test(
@@ -42,7 +42,7 @@ symlink_test(
 symlink_test(
     'usa-test',
     '/article.html?longname=United+States+of+America',
-    '/article.html?longname=USA&s=3t2X5xvsKo',
+    '/article.html?longname=USA',
 )
 
 symlink_test('usa-comparison', '/comparison.html?longnames=%5B%22United+States+of+America%22%2C%22Canada%22%5D', '/comparison.html?longnames=%5B%22USA%22%2C%22Canada%22%5D')
