+++ react/test/test_utils.ts
@@ -13,6 +13,7 @@ export const IS_TESTING = true
 export function comparison_page(locations: string[]): string {
     const params = new URLSearchParams()
     params.set('longnames', JSON.stringify(locations))
+    params.set('s', '3t2X5xvsKo')
     return `${TARGET}/comparison.html?${params.toString()}`
 }
